WIP terribly broken start of refactoring
