+++ build.gradle
@@ -3,7 +3,7 @@ plugins {
 }
 
 group = 'org.dustyroom'
-version = '0.4.1'
+version = '0.4.2'
 
 repositories {
     mavenCentral()
