+++ gradle/wrapper/gradle-wrapper.properties
@@ -1,6 +1,7 @@
-#Sun Feb 18 00:08:29 EET 2024
 distributionBase=GRADLE_USER_HOME
 distributionPath=wrapper/dists
 distributionUrl=https\://services.gradle.org/distributions/gradle-8.4-bin.zip
+networkTimeout=10000
+validateDistributionUrl=true
 zipStoreBase=GRADLE_USER_HOME
 zipStorePath=wrapper/dists
