+++ src/main/java/org/dustyroom/ui/components/OptionsMenuListener.java
@@ -1,0 +1,6 @@
+package org.dustyroom.ui.components;
+
+public interface OptionsMenuListener {
+
+    void toggleFullscreen();
+}
